+++ notebooks/promote_topics.py
@@ -21,7 +21,7 @@
 
 # COMMAND ----------
 
-# MAGIC %pip install /Volumes/<catalog>/<schema>/wheels/wikibricks-0.3.3-py3-none-any.whl
+# MAGIC %pip install /Volumes/<catalog>/<schema>/wheels/wikibricks-0.3.4-py3-none-any.whl
 # MAGIC %restart_python
 
 # COMMAND ----------